+++ packages/rax-scripts/src/config/pwa/server/webpack.config.base.js
@@ -3,7 +3,7 @@
 const webpack = require('webpack');
 const webpackMerge = require('webpack-merge');
 const webpackConfigBase = require('../../webapp/webpack.config.base');
-const getEntries = require('../../../utils/getEntries');
+const getEntries = require('../../../utils/getPWAEntries');
 
 const webpackConfig = webpackMerge(webpackConfigBase, {
   target: 'node',
